a2server.svg: Add a fill shape
I sometimes get asked: Why a fill shape and not a stroke?  A stroke of 1
here would give the same outcome for this shape.  And it would.  The
reason I don't do it however is that this gives me more control.  It
also lets me do what I'm going to do next more easily.  May be a better
way to do it, but this is what I'm going to do.