fb-group-cover.svg: Repair metadata
All right, while the contents of fb-group-cover.svg are probably not
subject to Copyright given their simplicity, the composite image is, and
the public domain license mark was carrying through to the composite
image.  Let's fix that!

Also noticed I had neglected to change the title in the boilerplate file
I'd copied when I started this SVG.  Fixed in both the boilerplate and
this file.